refactor: reorganize query helpers into modules
