+++ community_version/common/bm25.py
@@ -1,14 +1,18 @@
 """BM25-based re-ranking utilities for OpenSearch hits."""
 from __future__ import annotations
 
-from typing import Any, Dict, List, Tuple
+import json
+from typing import Any, Dict, List, Optional, Tuple
 
 import bm25s
+from opensearchpy.exceptions import NotFoundError
 
 try:
     import Stemmer  # type: ignore
 except ImportError:  # pragma: no cover - optional dependency
     Stemmer = None  # type: ignore
+
+from .models import RetrievalHit
 
 
 _BM25_STOPWORDS = "en"
@@ -123,4 +127,271 @@ def rerank_hits_with_bm25(
     return keep_long, keep_hot, combined
 
 
-__all__ = ["rerank_hits_with_bm25"]
+def _entity_should_clauses(entities: List[str]) -> List[Dict[str, Any]]:
+    should: List[Dict[str, Any]] = []
+    for ent in entities:
+        ent_l = ent.strip().lower()
+        if not ent_l:
+            continue
+        should.append({"term": {"explicit_terms": {"value": ent_l, "boost": 10.0}}})
+        should.append({"match_phrase": {"explicit_terms_text": {"query": ent_l, "boost": 8.0}}})
+        should.append({"match_phrase": {"content": {"query": ent, "boost": 3.0}}})
+    return should
+
+
+def _build_bm25_doc_query(question: str, entities: List[str], *, k: int) -> Dict[str, Any]:
+    should = _entity_should_clauses(entities)
+    return {
+        "size": k,
+        "_source": ["filepath", "category", "explicit_terms", "explicit_terms_text", "content"],
+        "query": {
+            "bool": {
+                "must": [{
+                    "multi_match": {
+                        "query": question,
+                        "fields": ["explicit_terms_text^6", "content^2", "category^0.5"],
+                        "type": "best_fields",
+                        "operator": "or",
+                    }
+                }],
+                "should": should,
+                "minimum_should_match": 1 if should else 0,
+            }
+        },
+    }
+
+
+def _build_bm25_chunk_query(
+    question: str,
+    entities: List[str],
+    *,
+    k: int,
+    anchor_paths: Optional[List[str]] = None,
+    strict: bool = True,
+) -> Dict[str, Any]:
+    should = _entity_should_clauses(entities)
+
+    operator = "and" if strict else "or"
+    msm = None if strict else "30%"
+
+    filters: List[Dict[str, Any]] = []
+    if anchor_paths:
+        filters.append({
+            "bool": {
+                "should": [
+                    {"terms": {"parent_filepath": anchor_paths}},
+                    {"terms": {"filepath": anchor_paths}},
+                ],
+                "minimum_should_match": 1,
+            }
+        })
+
+    mm: Dict[str, Any] = {
+        "query": question,
+        "fields": ["explicit_terms_text^8", "content^2", "category^0.5"],
+        "type": "best_fields",
+        "operator": operator,
+    }
+    if msm:
+        mm["minimum_should_match"] = msm
+
+    return {
+        "size": k,
+        "_source": [
+            "filepath", "parent_filepath", "chunk_index", "chunk_count",
+            "category", "explicit_terms", "explicit_terms_text", "content",
+        ],
+        "query": {
+            "bool": {
+                "filter": filters,
+                "must": [{"multi_match": mm}],
+                "should": should,
+                "minimum_should_match": 1 if should else 0,
+            }
+        },
+    }
+
+
+def bm25_retrieve_doc_anchors(
+    bm25_client: Any,
+    index: str,
+    *,
+    question: str,
+    entities: List[str],
+    k: int,
+    observability: bool,
+) -> Tuple[List[str], Dict[str, Any], List[Dict[str, Any]]]:
+    q = _build_bm25_doc_query(question, entities, k=k)
+    if observability:
+        print("\n[BM25_DOC_QUERY]\n" + json.dumps(q, ensure_ascii=False, indent=2, sort_keys=True, default=str))
+
+    res = bm25_client.search(index=index, body=q)
+    hits = res.get("hits", {}).get("hits", []) or []
+
+    paths: List[str] = []
+    for h in hits:
+        src = h.get("_source", {}) or {}
+        fp = src.get("filepath")
+        if fp:
+            paths.append(fp)
+
+    # unique preserve order
+    seen = set()
+    out: List[str] = []
+    for p in paths:
+        if p in seen:
+            continue
+        seen.add(p)
+        out.append(p)
+    return out, q, hits
+
+
+def bm25_retrieve_chunks(
+    bm25_client: Any,
+    index: str,
+    *,
+    question: str,
+    entities: List[str],
+    k: int,
+    anchor_paths: Optional[List[str]],
+    neighbor_window: int,
+    observability: bool,
+) -> Tuple[List[RetrievalHit], Dict[str, Any], List[Dict[str, Any]]]:
+    attempts: List[Tuple[Optional[List[str]], bool]] = []
+    if anchor_paths:
+        attempts.extend([
+            (anchor_paths, True),
+            (anchor_paths, False),
+            (None, True),
+            (None, False),
+        ])
+    else:
+        attempts.extend([(None, True), (None, False)])
+
+    q: Dict[str, Any] = {}
+    raw_hits: List[Dict[str, Any]] = []
+    for ap, strict in attempts:
+        q = _build_bm25_chunk_query(question, entities, k=k, anchor_paths=ap, strict=strict)
+        if observability:
+            print("\n[BM25_CHUNK_QUERY]\n" + json.dumps({**q, "_note": f"anchor={bool(ap)} strict={strict}"}, ensure_ascii=False, indent=2, sort_keys=True, default=str))
+        res = bm25_client.search(index=index, body=q)
+        raw_hits = res.get("hits", {}).get("hits", []) or []
+        if raw_hits:
+            break
+
+    ent_set = {e.strip().lower() for e in entities if e.strip()}
+    hits: List[RetrievalHit] = []
+    for i, h in enumerate(raw_hits, start=1):
+        src = h.get("_source", {}) or {}
+        explicit_terms = src.get("explicit_terms") or []
+        overlap = None
+        if ent_set and explicit_terms:
+            overlap = len({t.strip().lower() for t in explicit_terms} & ent_set)
+
+        path = src.get("parent_filepath") or src.get("filepath") or ""
+        hits.append(
+            RetrievalHit(
+                channel="bm25_chunk",
+                handle=f"B{i}",
+                index=index,
+                os_id=h.get("_id", ""),
+                score=float(h.get("_score") or 0.0),
+                path=path,
+                category=src.get("category") or "",
+                chunk_index=src.get("chunk_index"),
+                chunk_count=src.get("chunk_count"),
+                text=(src.get("content") or "").strip(),
+                explicit_terms=explicit_terms,
+                entity_overlap=overlap,
+            )
+        )
+
+    if neighbor_window > 0 and hits:
+        expanded = _expand_bm25_neighbors(
+            bm25_client,
+            index=index,
+            seed_hits=hits,
+            window=neighbor_window,
+            observability=observability,
+        )
+        # Re-handle
+        hits = [
+            RetrievalHit(**{**h.to_jsonable(), "handle": f"B{i+1}"}) for i, h in enumerate(expanded)
+        ]
+
+    return hits, q, raw_hits
+
+
+def _expand_bm25_neighbors(
+    bm25_client: Any,
+    *,
+    index: str,
+    seed_hits: List[RetrievalHit],
+    window: int,
+    observability: bool,
+) -> List[RetrievalHit]:
+    # group seed positions per doc path
+    doc_positions: Dict[str, List[int]] = {}
+    seed_score: Dict[str, float] = {}
+    for h in seed_hits:
+        seed_score[h.os_id] = h.score
+        if h.chunk_index is None:
+            continue
+        doc_positions.setdefault(h.path, []).append(int(h.chunk_index))
+
+    expanded: List[RetrievalHit] = []
+    for path, positions in doc_positions.items():
+        idxs: set[int] = set()
+        for pos in positions:
+            for off in range(-window, window + 1):
+                if pos + off >= 0:
+                    idxs.add(pos + off)
+        for ci in sorted(idxs):
+            chunk_id = f"{path}::chunk-{ci:03d}"
+            try:
+                doc = bm25_client.get(index=index, id=chunk_id)
+            except NotFoundError:
+                continue
+            src = doc.get("_source", {}) or {}
+            expanded.append(
+                RetrievalHit(
+                    channel="bm25_chunk" if chunk_id in seed_score else "bm25_neighbor",
+                    handle="B0",  # temporary, reassigned later
+                    index=index,
+                    os_id=chunk_id,
+                    score=float(seed_score.get(chunk_id, 0.0)),
+                    path=src.get("parent_filepath") or src.get("filepath") or path,
+                    category=src.get("category") or "",
+                    chunk_index=src.get("chunk_index"),
+                    chunk_count=src.get("chunk_count"),
+                    text=(src.get("content") or "").strip(),
+                    explicit_terms=src.get("explicit_terms") or [],
+                    entity_overlap=None,
+                )
+            )
+
+    # de-dup (path, chunk_index) preserve order
+    seen: set[Tuple[str, Optional[int]]] = set()
+    uniq: List[RetrievalHit] = []
+    for h in expanded:
+        key = (h.path, h.chunk_index)
+        if key in seen:
+            continue
+        seen.add(key)
+        uniq.append(h)
+
+    if observability:
+        print(f"\n[BM25_NEIGHBOR_EXPANSION] window={window} expanded_chunks={len(uniq)}")
+
+    return uniq
+
+
+__all__ = [
+    "rerank_hits_with_bm25",
+    "_entity_should_clauses",
+    "_build_bm25_doc_query",
+    "_build_bm25_chunk_query",
+    "bm25_retrieve_doc_anchors",
+    "bm25_retrieve_chunks",
+    "_expand_bm25_neighbors",
+]
