+++ community_version/query.py
@@ -24,46 +24,26 @@ import argparse
 import json
 import re
 import time
-from dataclasses import asdict, dataclass
 from typing import Any, Dict, List, Optional, Sequence, Tuple
 
-from opensearchpy.exceptions import NotFoundError
-
-from common.embeddings import EmbeddingModel, to_list
-from common.llm import load_llm
+from common.bm25 import bm25_retrieve_chunks, bm25_retrieve_doc_anchors
+from common.embeddings import vector_retrieve_chunks
+from common.llm import (
+    build_grounding_prompt,
+    build_refine_prompt,
+    build_single_pass_prompt,
+    build_vector_only_prompt,
+    call_llm_chat,
+    load_llm,
+)
 from common.logging import get_logger
-from common.named_entity import normalize_entities, post_ner
+from common.models import RetrievalHit
+from common.named_entity import extract_entities
 from common.opensearch_client import create_long_client, create_vector_client
 
 LOGGER = get_logger(__name__)
 
 _CITATION_RE = re.compile(r"\[(B\d+|V\d+)\]")
-
-
-# --------------------------------------------------------------------------------------
-# Data structures
-# --------------------------------------------------------------------------------------
-
-@dataclass(frozen=True)
-class RetrievalHit:
-    channel: str  # bm25_doc | bm25_chunk | bm25_neighbor | vector
-    handle: str   # B1..Bn or V1..Vn
-    index: str
-    os_id: str
-    score: float
-
-    path: str
-    category: str
-    chunk_index: Optional[int] = None
-    chunk_count: Optional[int] = None
-
-    text: str = ""
-
-    explicit_terms: Optional[List[str]] = None
-    entity_overlap: Optional[int] = None
-
-    def to_jsonable(self) -> Dict[str, Any]:
-        return asdict(self)
 
 
 # --------------------------------------------------------------------------------------
@@ -103,600 +83,8 @@ def parse_args(argv: Sequence[str] | None = None) -> argparse.Namespace:
     return p.parse_args(argv)
 
 
-def _json(obj: Any) -> str:
-    return json.dumps(obj, ensure_ascii=False, indent=2, sort_keys=True, default=str)
-
-
 def _extract_citations(answer: str) -> List[str]:
     return sorted(set(m.group(1) for m in _CITATION_RE.finditer(answer)))
-
-
-def _messages_to_prompt(messages: List[Dict[str, str]]) -> str:
-    """Fallback formatting for non-chat completion interfaces."""
-    parts: List[str] = []
-    for m in messages:
-        role = (m.get("role") or "user").upper()
-        content = m.get("content") or ""
-        parts.append(f"{role}:\n{content}".strip())
-    return "\n\n".join(parts).strip()
-
-
-# --------------------------------------------------------------------------------------
-# Entity extraction (no LLM policy)
-# --------------------------------------------------------------------------------------
-
-def extract_entities(question: str) -> List[str]:
-    try:
-        ner_payload = post_ner(question)
-        return normalize_entities(ner_payload)
-    except Exception as exc:
-        LOGGER.warning("NER failed (%s). Proceeding without entities.", exc)
-        return []
-
-
-def _entity_should_clauses(entities: List[str]) -> List[Dict[str, Any]]:
-    should: List[Dict[str, Any]] = []
-    for ent in entities:
-        ent_l = ent.strip().lower()
-        if not ent_l:
-            continue
-        should.append({"term": {"explicit_terms": {"value": ent_l, "boost": 10.0}}})
-        should.append({"match_phrase": {"explicit_terms_text": {"query": ent_l, "boost": 8.0}}})
-        should.append({"match_phrase": {"content": {"query": ent, "boost": 3.0}}})
-    return should
-
-
-# --------------------------------------------------------------------------------------
-# OpenSearch queries
-# --------------------------------------------------------------------------------------
-
-def _build_bm25_doc_query(question: str, entities: List[str], *, k: int) -> Dict[str, Any]:
-    should = _entity_should_clauses(entities)
-    return {
-        "size": k,
-        "_source": ["filepath", "category", "explicit_terms", "explicit_terms_text", "content"],
-        "query": {
-            "bool": {
-                "must": [{
-                    "multi_match": {
-                        "query": question,
-                        "fields": ["explicit_terms_text^6", "content^2", "category^0.5"],
-                        "type": "best_fields",
-                        "operator": "or",
-                    }
-                }],
-                "should": should,
-                "minimum_should_match": 1 if should else 0,
-            }
-        },
-    }
-
-
-def _build_bm25_chunk_query(
-    question: str,
-    entities: List[str],
-    *,
-    k: int,
-    anchor_paths: Optional[List[str]] = None,
-    strict: bool = True,
-) -> Dict[str, Any]:
-    should = _entity_should_clauses(entities)
-
-    operator = "and" if strict else "or"
-    msm = None if strict else "30%"
-
-    filters: List[Dict[str, Any]] = []
-    if anchor_paths:
-        filters.append({
-            "bool": {
-                "should": [
-                    {"terms": {"parent_filepath": anchor_paths}},
-                    {"terms": {"filepath": anchor_paths}},
-                ],
-                "minimum_should_match": 1,
-            }
-        })
-
-    mm: Dict[str, Any] = {
-        "query": question,
-        "fields": ["explicit_terms_text^8", "content^2", "category^0.5"],
-        "type": "best_fields",
-        "operator": operator,
-    }
-    if msm:
-        mm["minimum_should_match"] = msm
-
-    return {
-        "size": k,
-        "_source": [
-            "filepath", "parent_filepath", "chunk_index", "chunk_count",
-            "category", "explicit_terms", "explicit_terms_text", "content",
-        ],
-        "query": {
-            "bool": {
-                "filter": filters,
-                "must": [{"multi_match": mm}],
-                "should": should,
-                "minimum_should_match": 1 if should else 0,
-            }
-        },
-    }
-
-
-def _build_vector_query(
-    query_vector: List[float],
-    *,
-    k: int,
-    candidate_k: int,
-    anchor_paths: Optional[List[str]] = None,
-    vector_field: str = "embedding",
-) -> Dict[str, Any]:
-    knn_clause = {"knn": {vector_field: {"vector": query_vector, "k": candidate_k}}}
-    return {
-        "size": k,
-        "_source": ["path", "category", "chunk_index", "chunk_count", "text"],
-        "query": {
-            "bool": {
-                "must": [knn_clause],
-                "filter": [{"terms": {"path": anchor_paths}}] if anchor_paths else [],
-            }
-        },
-    }
-
-
-# --------------------------------------------------------------------------------------
-# Retrieval
-# --------------------------------------------------------------------------------------
-
-def bm25_retrieve_doc_anchors(
-    bm25_client: Any,
-    index: str,
-    *,
-    question: str,
-    entities: List[str],
-    k: int,
-    observability: bool,
-) -> Tuple[List[str], Dict[str, Any], List[Dict[str, Any]]]:
-    q = _build_bm25_doc_query(question, entities, k=k)
-    if observability:
-        print("\n[BM25_DOC_QUERY]\n" + _json(q))
-
-    res = bm25_client.search(index=index, body=q)
-    hits = res.get("hits", {}).get("hits", []) or []
-
-    paths: List[str] = []
-    for h in hits:
-        src = h.get("_source", {}) or {}
-        fp = src.get("filepath")
-        if fp:
-            paths.append(fp)
-
-    # unique preserve order
-    seen = set()
-    out: List[str] = []
-    for p in paths:
-        if p in seen:
-            continue
-        seen.add(p)
-        out.append(p)
-    return out, q, hits
-
-
-def bm25_retrieve_chunks(
-    bm25_client: Any,
-    index: str,
-    *,
-    question: str,
-    entities: List[str],
-    k: int,
-    anchor_paths: Optional[List[str]],
-    neighbor_window: int,
-    observability: bool,
-) -> Tuple[List[RetrievalHit], Dict[str, Any], List[Dict[str, Any]]]:
-    attempts: List[Tuple[Optional[List[str]], bool]] = []
-    if anchor_paths:
-        attempts.extend([
-            (anchor_paths, True),
-            (anchor_paths, False),
-            (None, True),
-            (None, False),
-        ])
-    else:
-        attempts.extend([(None, True), (None, False)])
-
-    q: Dict[str, Any] = {}
-    raw_hits: List[Dict[str, Any]] = []
-    for ap, strict in attempts:
-        q = _build_bm25_chunk_query(question, entities, k=k, anchor_paths=ap, strict=strict)
-        if observability:
-            print("\n[BM25_CHUNK_QUERY]\n" + _json({**q, "_note": f"anchor={bool(ap)} strict={strict}"}))
-        res = bm25_client.search(index=index, body=q)
-        raw_hits = res.get("hits", {}).get("hits", []) or []
-        if raw_hits:
-            break
-
-    ent_set = {e.strip().lower() for e in entities if e.strip()}
-    hits: List[RetrievalHit] = []
-    for i, h in enumerate(raw_hits, start=1):
-        src = h.get("_source", {}) or {}
-        explicit_terms = src.get("explicit_terms") or []
-        overlap = None
-        if ent_set and explicit_terms:
-            overlap = len({t.strip().lower() for t in explicit_terms} & ent_set)
-
-        path = src.get("parent_filepath") or src.get("filepath") or ""
-        hits.append(
-            RetrievalHit(
-                channel="bm25_chunk",
-                handle=f"B{i}",
-                index=index,
-                os_id=h.get("_id", ""),
-                score=float(h.get("_score") or 0.0),
-                path=path,
-                category=src.get("category") or "",
-                chunk_index=src.get("chunk_index"),
-                chunk_count=src.get("chunk_count"),
-                text=(src.get("content") or "").strip(),
-                explicit_terms=explicit_terms,
-                entity_overlap=overlap,
-            )
-        )
-
-    if neighbor_window > 0 and hits:
-        expanded = _expand_bm25_neighbors(
-            bm25_client,
-            index=index,
-            seed_hits=hits,
-            window=neighbor_window,
-            observability=observability,
-        )
-        # Re-handle
-        hits = [
-            RetrievalHit(**{**h.to_jsonable(), "handle": f"B{i+1}"}) for i, h in enumerate(expanded)
-        ]
-
-    return hits, q, raw_hits
-
-
-def _expand_bm25_neighbors(
-    bm25_client: Any,
-    *,
-    index: str,
-    seed_hits: List[RetrievalHit],
-    window: int,
-    observability: bool,
-) -> List[RetrievalHit]:
-    # group seed positions per doc path
-    doc_positions: Dict[str, List[int]] = {}
-    seed_score: Dict[str, float] = {}
-    for h in seed_hits:
-        seed_score[h.os_id] = h.score
-        if h.chunk_index is None:
-            continue
-        doc_positions.setdefault(h.path, []).append(int(h.chunk_index))
-
-    expanded: List[RetrievalHit] = []
-    for path, positions in doc_positions.items():
-        idxs: set[int] = set()
-        for pos in positions:
-            for off in range(-window, window + 1):
-                if pos + off >= 0:
-                    idxs.add(pos + off)
-        for ci in sorted(idxs):
-            chunk_id = f"{path}::chunk-{ci:03d}"
-            try:
-                doc = bm25_client.get(index=index, id=chunk_id)
-            except NotFoundError:
-                continue
-            src = doc.get("_source", {}) or {}
-            expanded.append(
-                RetrievalHit(
-                    channel="bm25_chunk" if chunk_id in seed_score else "bm25_neighbor",
-                    handle="B0",  # temporary, reassigned later
-                    index=index,
-                    os_id=chunk_id,
-                    score=float(seed_score.get(chunk_id, 0.0)),
-                    path=src.get("parent_filepath") or src.get("filepath") or path,
-                    category=src.get("category") or "",
-                    chunk_index=src.get("chunk_index"),
-                    chunk_count=src.get("chunk_count"),
-                    text=(src.get("content") or "").strip(),
-                    explicit_terms=src.get("explicit_terms") or [],
-                    entity_overlap=None,
-                )
-            )
-
-    # de-dup (path, chunk_index) preserve order
-    seen: set[Tuple[str, Optional[int]]] = set()
-    uniq: List[RetrievalHit] = []
-    for h in expanded:
-        key = (h.path, h.chunk_index)
-        if key in seen:
-            continue
-        seen.add(key)
-        uniq.append(h)
-
-    if observability:
-        print(f"\n[BM25_NEIGHBOR_EXPANSION] window={window} expanded_chunks={len(uniq)}")
-
-    return uniq
-
-
-def vector_retrieve_chunks(
-    vec_client: Any,
-    index: str,
-    *,
-    question: str,
-    anchor_paths: Optional[List[str]],
-    k: int,
-    candidate_k: int,
-    observability: bool,
-    vector_field: str = "embedding",
-) -> Tuple[List[RetrievalHit], Dict[str, Any], List[Dict[str, Any]]]:
-    embedder = EmbeddingModel()
-    qvec = to_list(embedder.encode([question])[0])
-
-    q = _build_vector_query(qvec, k=k, candidate_k=candidate_k, anchor_paths=anchor_paths, vector_field=vector_field)
-
-    if observability:
-        dbg = json.loads(json.dumps(q))
-        try:
-            dbg["query"]["bool"]["must"][0]["knn"][vector_field]["vector"] = "<omitted>"
-        except Exception:
-            pass
-        print("\n[VECTOR_QUERY]\n" + _json(dbg))
-
-    res = vec_client.search(index=index, body=q)
-    raw_hits = res.get("hits", {}).get("hits", []) or []
-
-    hits: List[RetrievalHit] = []
-    for i, h in enumerate(raw_hits, start=1):
-        src = h.get("_source", {}) or {}
-        hits.append(
-            RetrievalHit(
-                channel="vector",
-                handle=f"V{i}",
-                index=index,
-                os_id=h.get("_id", ""),
-                score=float(h.get("_score") or 0.0),
-                path=src.get("path") or "",
-                category=src.get("category") or "",
-                chunk_index=src.get("chunk_index"),
-                chunk_count=src.get("chunk_count"),
-                text=(src.get("text") or "").strip(),
-            )
-        )
-
-    return hits, q, raw_hits
-
-
-# --------------------------------------------------------------------------------------
-# LLM call + response normalization
-# --------------------------------------------------------------------------------------
-
-def _extract_llm_text(resp: Any) -> str:
-    """Normalize many possible completion response shapes into a plain string."""
-    if resp is None:
-        return ""
-
-    if isinstance(resp, str):
-        return resp.strip()
-
-    # llama_cpp and many other clients return dicts
-    if isinstance(resp, dict):
-        choices = resp.get("choices")
-        if isinstance(choices, list) and choices:
-            c0 = choices[0]
-            if isinstance(c0, dict):
-                # chat completion style
-                msg = c0.get("message")
-                if isinstance(msg, dict) and "content" in msg:
-                    return str(msg.get("content") or "").strip()
-                # text completion style
-                if "text" in c0:
-                    return str(c0.get("text") or "").strip()
-                # streaming delta style
-                delta = c0.get("delta")
-                if isinstance(delta, dict) and "content" in delta:
-                    return str(delta.get("content") or "").strip()
-        # some wrappers return {"content": "..."}
-        if "content" in resp and isinstance(resp["content"], str):
-            return resp["content"].strip()
-        return str(resp).strip()
-
-    # OpenAI v1 python client response objects have .choices
-    choices = getattr(resp, "choices", None)
-    if choices and isinstance(choices, list):
-        c0 = choices[0]
-        # chat
-        msg = getattr(c0, "message", None)
-        if msg is not None:
-            content = getattr(msg, "content", None)
-            if isinstance(content, str):
-                return content.strip()
-        # completion
-        text = getattr(c0, "text", None)
-        if isinstance(text, str):
-            return text.strip()
-
-    # Fallback: stringify
-    return str(resp).strip()
-
-
-def call_llm_chat(
-    llm: Any,
-    *,
-    messages: List[Dict[str, str]],
-    model: Optional[str],
-    temperature: float,
-    top_p: float,
-    max_tokens: int,
-) -> str:
-    """Chat-style completion, with explicit support for llama_cpp.Llama.
-
-    This fixes the failure mode where calling a llama_cpp model directly
-    (llm(prompt)) returns a raw dict completion object rather than the text.
-    """
-    # 1) llama_cpp: use create_chat_completion (best behavior for role-based prompts)
-    create_chat = getattr(llm, "create_chat_completion", None)
-    if callable(create_chat):
-        resp = create_chat(
-            messages=messages,
-            temperature=temperature,
-            top_p=top_p,
-            max_tokens=max_tokens,
-        )
-        return _extract_llm_text(resp)
-
-    # 2) OpenAI-like client: llm.chat.completions.create(...)
-    chat = getattr(llm, "chat", None)
-    if chat is not None:
-        completions = getattr(chat, "completions", None)
-        create = getattr(completions, "create", None) if completions is not None else None
-        if callable(create):
-            kwargs: Dict[str, Any] = {
-                "messages": messages,
-                "temperature": temperature,
-                "max_tokens": max_tokens,
-            }
-            if model:
-                kwargs["model"] = model
-            resp = create(**kwargs)
-            return _extract_llm_text(resp)
-
-    # 3) Some wrappers provide invoke(messages) -> str/dict
-    invoke = getattr(llm, "invoke", None)
-    if callable(invoke):
-        resp = invoke(messages)
-        return _extract_llm_text(resp)
-
-    # 4) Text completion fallback: convert messages to prompt and try create_completion / __call__
-    prompt = _messages_to_prompt(messages)
-
-    create_completion = getattr(llm, "create_completion", None)
-    if callable(create_completion):
-        resp = create_completion(prompt=prompt, temperature=temperature, top_p=top_p, max_tokens=max_tokens)
-        return _extract_llm_text(resp)
-
-    if callable(llm):
-        # llama_cpp.__call__ uses a low default max_tokens; pass explicit args if supported
-        try:
-            resp = llm(prompt, temperature=temperature, top_p=top_p, max_tokens=max_tokens)
-        except TypeError:
-            resp = llm(prompt)
-        return _extract_llm_text(resp)
-
-    raise TypeError("Unsupported LLM client type returned by load_llm().")
-
-
-# --------------------------------------------------------------------------------------
-# Prompt building (explicit separation)
-# --------------------------------------------------------------------------------------
-
-def _format_hits(hits: List[RetrievalHit], *, title: str) -> str:
-    blocks = [f"## {title}"]
-    for h in hits:
-        meta = [f"handle={h.handle}", f"score={h.score:.4f}", f"path={h.path}"]
-        if h.chunk_index is not None and h.chunk_count is not None:
-            meta.append(f"chunk={h.chunk_index}/{h.chunk_count}")
-        if h.category:
-            meta.append(f"category={h.category}")
-        blocks.append(f"[{h.handle}] ({', '.join(meta)})\n{h.text}".strip())
-    return "\n\n".join(blocks).strip()
-
-
-def build_grounding_prompt(question: str, bm25_hits: List[RetrievalHit]) -> List[Dict[str, str]]:
-    print("***** Building grounding prompt...")
-
-    context = _format_hits(bm25_hits, title="BM25 Grounding Evidence (authoritative facts)")
-    system = (
-        "Answer using ONLY the BM25 Grounding Evidence.\n"
-        "Rules:\n"
-        "- Every factual claim must cite at least one [B#].\n"
-        "- If evidence does not support the answer, respond: I don't know based on the provided evidence.\n"
-        "- Do not quote the evidence headers/metadata. Use them only for citations.\n"
-        "- If there is conflicting evidence, disclose the conflict and data.\n"
-    )
-    user = f"QUESTION:\n{question}\n\nGROUNDING_EVIDENCE:\n{context}\n"
-
-    prompt_details = [{"role": "system", "content": system}, {"role": "user", "content": user}]
-
-    print("=====================================================")
-    print(_messages_to_prompt(prompt_details))
-    print("=====================================================")
-
-    return prompt_details
-
-
-def build_vector_only_prompt(question: str, vec_hits: List[RetrievalHit]) -> List[Dict[str, str]]:
-    print("***** Building vector-only prompt...")
-    context = _format_hits(vec_hits, title="Vector Evidence (semantic fallback)")
-    system = (
-        "Answer using ONLY the Vector Evidence.\n"
-        "Rules:\n"
-        "- Every factual claim must cite at least one [V#].\n"
-        "- If evidence does not support the answer, respond: I don't know based on the provided evidence.\n"
-        "- Do not quote the evidence headers/metadata.\n"
-    )
-    user = f"QUESTION:\n{question}\n\nEVIDENCE:\n{context}\n"
-    prompt_details = [{"role": "system", "content": system}, {"role": "user", "content": user}]
-
-    print("=====================================================")
-    print(_messages_to_prompt(prompt_details))
-    print("=====================================================")
-
-    return prompt_details
-
-
-def build_refine_prompt(question: str, grounded_draft: str, vec_hits: List[RetrievalHit]) -> List[Dict[str, str]]:
-    print("***** Building refine prompt...")
-
-    vec_context = _format_hits(vec_hits, title="Vector Semantic Context (phrasing/terminology support)")
-    system = (
-        "Rewrite the grounded draft for clarity.\n"
-        "Rules:\n"
-        "- Do NOT add new factual claims beyond what appears in the draft.\n"
-        "- Preserve [B#] citations exactly.\n"
-        "- You may add brief non-factual clarifications supported by vector context and cite [V#].\n"
-        "- Output ONLY the rewritten answer.\n"
-        "- If there is conflicting evidence, disclose the conflict and data.\n"
-    )
-    user = (
-        f"QUESTION:\n{question}\n\n"
-        f"GROUNDED_DRAFT:\n{grounded_draft}\n\n"
-        f"{vec_context}\n"
-    )
-    prompt_details = [{"role": "system", "content": system}, {"role": "user", "content": user}]
-
-    print("=====================================================")
-    print(_messages_to_prompt(prompt_details))
-    print("=====================================================")
-
-    return prompt_details
-
-
-def build_single_pass_prompt(question: str, bm25_hits: List[RetrievalHit], vec_hits: List[RetrievalHit]) -> List[Dict[str, str]]:
-    print("***** Building single-pass prompt...")
-
-    bm25_context = _format_hits(bm25_hits, title="BM25 Grounding Evidence (authoritative facts)")
-    vec_context = _format_hits(vec_hits, title="Vector Semantic Context (phrasing/terminology support)")
-    system = (
-        "Answer using the provided contexts.\n"
-        "Separation of concerns:\n"
-        "- Use BM25 Grounding Evidence for factual claims.\n"
-        "- Use Vector Semantic Context only for wording/terminology, not new facts.\n"
-        "Rules:\n"
-        "- Every factual claim must cite [B#].\n"
-        "- If BM25 evidence does not support the answer, respond: I don't know based on the provided evidence.\n"
-    )
-    user = f"QUESTION:\n{question}\n\n{bm25_context}\n\n{vec_context}\n"
-    prompt_details = [{"role": "system", "content": system}, {"role": "user", "content": user}]
-
-    print("=====================================================")
-    print(_messages_to_prompt(prompt_details))
-    print("=====================================================")
-
-    return prompt_details
 
 
 # --------------------------------------------------------------------------------------
